// store/actions/types/app.ts
export enum GetUsersCountsTypes {
  GET_USERS_COUNTS_LOADING = "GET_USERS_COUNTS_LOADING",
  GET_USERS_COUNTS_SUCCESS = "GET_USERS_COUNTS_SUCCESS",
  GET_USERS_COUNTS_ERROR = "GET_USERS_COUNTS_ERROR",
}

export enum GetReservationsCountsTypes {
  GET_RESERVATIONS_COUNTS_LOADING = "GET_RESERVATIONS_COUNTS_LOADING",
  GET_RESERVATIONS_COUNTS_SUCCESS = "GET_RESERVATIONS_COUNTS_SUCCESS",
  GET_RESERVATIONS_COUNTS_ERROR = "GET_RESERVATIONS_COUNTS_ERROR",
}

export enum GetUsersJoiningTypes {
  GET_USERS_JOINING_LOADING = "GET_USERS_JOINING_LOADING",
  GET_USERS_JOINING_SUCCESS = "GET_USERS_JOINING_SUCCESS",
  GET_USERS_JOINING_ERROR = "GET_USERS_JOINING_ERROR",
}

export enum GetPatientsCitiesTypes {
  GET_PATIENTS_CITIES_LOADING = "GET_PATIENTS_CITIES_LOADING",
  GET_PATIENTS_CITIES_SUCCESS = "GET_PATIENTS_CITIES_SUCCESS",
  GET_PATIENTS_CITIES_ERROR = "GET_PATIENTS_CITIES_ERROR",
}

export enum GetDoctorsCitiesTypes {
  GET_DOCTORS_CITIES_LOADING = "GET_DOCTORS_CITIES_LOADING",
  GET_DOCTORS_CITIES_SUCCESS = "GET_DOCTORS_CITIES_SUCCESS",
  GET_DOCTORS_CITIES_ERROR = "GET_DOCTORS_CITIES_ERROR",
}